requirements.txt
aiohttp==3.3.0
alembic==0.9.9
amqp==2.3.2
aniso8601==3.0.2
async-timeout==3.0.0
attrs==18.1.0
Babel==2.6.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
billiard==3.5.0.4
blinker==1.4
celery==4.2.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
coverage==4.5.1
cssmin==0.2.0
defusedxml==0.5.0
eventlet==0.23.0
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Assets==0.12
Flask-BasicAuth==0.2.0
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Caching==1.4.0
Flask-Celery-Helper==1.1.0
Flask-DebugToolbar==0.10.1
Flask-Email==1.4.4
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
flask-mongoengine==0.9.5
Flask-OAuth==0.12
Flask-OpenID==1.2.5
Flask-Principal==0.4.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
flask-youku==0.1
flower==0.9.2
greenlet==0.4.13
HTMLParser==0.0.2
httplib2==0.11.3
idna==2.6
idna-ssl==1.0.1
itsdangerous==0.24
Jinja2==2.10
jsmin==2.2.2
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
mongoengine==0.15.0
multidict==4.3.1
mysql-connector==2.1.6
oauth2==1.9.0.post1
panda==0.3.1
Pillow==5.1.0
pycparser==2.18
pymongo==3.7.0
python-dateutil==2.7.3
python-editor==1.0.3
python3-openid==3.1.0
pytz==2018.5
redis==2.10.6
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.8
tornado==5.1
urllib3==1.22
vine==1.1.4
virtualenv==16.0.0
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.2.1
yarl==1.2.4
